Update CCScroller.cpp
Compared to android, the cocos2dx coordinate system has float value. So I don't think we should round the return value.

And the current code would cause a problem like this
=============
when I'm doing like this

m_scroller->startScroll(0.1818,0,100,0,1000);

then compute the offset.

At beginning the return value appears normal, but when approaching the end scroll, the getCurrX method would return 100.1818, which is larger than the finalX.

After I remove the round method, it seems to be normal now.